Api modifications in User Dashboard v4.7

+++ database/migrations/2018_06_11_203940_add_leave_pet_leave_money_and_pet_amount_to_tell_us_about_you.php
@@ -15,7 +15,7 @@ class AddLeavePetLeaveMoneyAndPetAmountToTellUsAboutYou extends Migration
     {
         Schema::table('tellUsAboutYou', function (Blueprint $table) {
             $table->boolean('leaveMoney')->nullable()->comment('required for pets field');
-            $table->integer('petAmount')->nullable()->comment('required for pets amount for pets care');
+            $table->float('petAmount')->nullable()->comment('required for pets amount for pets care');
         });
     }
